+++ versiondb/client/convert_to_sst.go
@@ -90,6 +90,9 @@ func convertSingleStore(store string, changeSetDir, sstDir string, sstFileSize u
 	if err != nil {
 		return err
 	}
+	if len(csFiles) == 0 {
+		return nil
+	}
 
 	prefix := []byte(fmt.Sprintf(tsrocksdb.StorePrefixTpl, store))
 	isEmpty := true
@@ -179,6 +182,10 @@ func scanChangeSetFiles(changeSetDir, store string) ([]FileWithVersion, error) {
 	storeDir := filepath.Join(changeSetDir, store)
 	entries, err := os.ReadDir(storeDir)
 	if err != nil {
+		// assume the change set files are taken from older versions, don't include all stores.
+		if os.IsNotExist(err) {
+			return nil, nil
+		}
 		return nil, err
 	}
 	fileNames := make([]string, len(entries))