+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [#986](https://github.com/crypto-org-chain/cronos/pull/986) Use go 1.20.
 - [#984](https://github.com/crypto-org-chain/cronos/pull/984) experimental integration of memiavl.
+- [#985](https://github.com/crypto-org-chain/cronos/pull/985) Fix versiondb verify command on older versions
 
 *April 13, 2023*